refactor(week-calendar): drop fractional hours; whole-hour blocks for now
The fractional-hours ("minutes") work is being split out for independent,
app-wide verification and re-applied on a stacked branch
(feat/week-calendar-minutes). This branch reverts to INTEGER, whole-hour
blocks so the calendar can ship on its own; block positioning (start/end
times) is unchanged.

Reverted to integer / whole-hour:
- TimeEntry.hours and WorkItem.{estimated,remaining,logged}_hours -> Integer.
- The int->float fanout across pulse, capacity_service, admin, workitems,
  dedupe (restored to main), and time_blocks (_hours_between now whole hours).
- database.py migration: keep the start/end columns + index, drop the NUMERIC
  widening. Standalone migrate script likewise adds columns only.
- Contract goldens + conftest back to integer seed.
- Calendar grid snaps to whole hours (stepMinutes 60); LogHours/time-block
  responses type hours as int; openapi regenerated (client types unchanged —
  TS maps int and float alike to number).

Kept: start_time/end_time positioning, and the whole calendar feature.

+++ backend/routers/admin.py
@@ -326,8 +326,8 @@ class EmployeeTicketResponse(BaseModel):
     project_name: str
     assigned_to: int | None
     assigned_to_name: str | None
-    estimated_hours: float | None
-    logged_hours: float
+    estimated_hours: int | None
+    logged_hours: int
     created_at: datetime
     updated_at: datetime
 
@@ -771,8 +771,8 @@ class WeeklyTicket(BaseModel):
     type: str
     priority: str
     assignee_name: str | None
-    estimated_hours: float | None
-    logged_hours: float | None
+    estimated_hours: int | None
+    logged_hours: int | None
     completed_at: str | None  # ISO datetime, only set for the `done_this_week` bucket
 
 
@@ -960,7 +960,7 @@ class TimeEntryRow(BaseModel):
     Developer joins so the frontend can render without nested lookups."""
 
     id: int
-    hours: float
+    hours: int
     description: str | None
     logged_at: datetime
 
@@ -986,7 +986,7 @@ class TimeEntriesResponse(BaseModel):
     frontend can warn when its filters return more than the cap."""
 
     rows: list[TimeEntryRow]
-    total_hours: float
+    total_hours: int
     total_rows: int
     truncated: bool
 
@@ -1072,7 +1072,7 @@ def list_time_entries(
     entries = fetched[:TIME_ENTRIES_MAX_ROWS]
 
     rows: list[TimeEntryRow] = []
-    total_hours = 0.0
+    total_hours = 0
     for te in entries:
         wi = te.work_item
         proj = wi.project if wi else None
